feat: migrate DuosHeader.spec
